Verilog Code for blinking lights
There are a few modules that go together to make a light blink one and off every second. This is done using different modules, thought it can be done in one, for the purpose of using different techniques. There is also a module that will use all 4 LEDs in different ways.